migrated reading to c++

+++ src/obis.c
@@ -177,7 +177,7 @@ int obis_lookup_alias(const char *alias, Obis *id) {
 	return ERR_NOT_FOUND;
 }
 
-int Obis::unparse(char *buffer, size_t n) {
+size_t Obis::unparse(char *buffer, size_t n) {
 	return snprintf(buffer, n, "%i-%i:%i.%i.%i*%i",
                   groups.media,
                   groups.channel,
